Extract shared media grid pieces from DiveMediaSection
